Add an option to stop make on various errors
By default, many of the processes in the webwml build are run with a
leading "-" so that make will ignore errors. Replace *most* of
those (*not* including rm calls for make clean) with $(IGNORE) and add
a conditional to Makefile.common to control what $(IGNORE) does.

By default, things will work just the same as before. If you set
STRICT_ERROR_CHECKS=1 in Makefile.common, the build will now be much
more strict. Most people can just ignore the change to add this
option, but it's helpful when debugging the scripts to have things
fail cleanly and quickly.