oval: fix encoding attribute
This replaces the 'ASCII' encoding attribute with 'UTF-8', which should
be correct.

Having 'ASCII' is problematic because the IANA name for ASCII encoding
is 'US-ASCII', and 'ASCII' is not an enumerated alias. Beyond that, it's
unclear why ElementTree is saying the encoding is 'ASCII', as the
documentation claims it should output 'US-ASCII' or 'UTF-8', as far as I
can tell. Perhaps it's the round-trip via string that's messing
something up.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>